+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2023-07-05
 
 - 新增显示块 ID 功能 | Add display block ID function.
+- 新增显示块内容功能 | Add display block content function.
 
 ## 2023-07-04